Support new format OpenSSH keys for all algorithms
