+++ test/com/trilead/ssh2/signature/dsa-testkey-unprotected-newformat.txt
@@ -1,0 +1,39 @@
+-----BEGIN OPENSSH PRIVATE KEY-----
+b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAADMQAAAAdz
+c2gtZHNzAAABAQCTwRCtqVzAb54+8dO5HnIpI6c/xxemJaTWSIkkkqT6cPQWnYMY
+T8q+ozxeuml39qWR7/EjJEGyzuCM5E3gtmEB2fJtqmDtrNS+3W28WBcsz/cRFr0O
+xohHOaTY6nOHnXTIwEUMclXaX3Fqc1erN20wGSuvqWiQuoSLo56KJ05/DmhNrFdf
+2jnS20k3SfKqCfj9wTUh9XFBRbH3aP/1iaDXjSogg/86N9bgwj+XkDclBySNOKp4
+xH6youAsXzaicMtwg2JHwLJbpVMOrtXdcJoUpq80mTvMj1asJWhuw/sCOzat5G92
+yE3ejJPMNK7Zxmpf98KDQDi5NtdOJCqqMq0jAAAAFQC6+Z5aw4Z+R71Hv3PsK8yz
+d5zdlQAAAQAK54Po+geyRR3uypcIw6Z5e9dkbsjYfeRYE3bHwAVR92/etX8uKIJU
+xIaBmzSYWqxqmtamxNe+VIU5MlE3FOA+XcqG+6ALS7Su6Eb3GVNG2W7YY6Oo8vEb
+IfU7yY3FWz5r69zjw20wMH/191NVxDQBWiwDLvpGLpbJQFNL58x2JFAiBglb6Jfs
+m3SCXJpg8wsOgGrpv8/sB/cO9ugV+/Kdv3DLHDJ8H/+qdPazwFatSywm1+DEyuZ4
+8ZQJGtcfrx7qQ8aEbiQGZCXDlIrrBY2/3OV3n774fvssc+GN98YpcVg3LOZW6rMp
+GaCHnLmZ9kupaSFjEMsyIq917H0/+HpTAAABADnMJBHQgFaxq/sG5ipM0Ct8cUk4
+Tsw4ijUFyuzWeaaeLs3q8N5bdZH/BHZoQS2L9zEoUTXWvhyc1X3tjVvDL+w6Hrek
+lzY6Wi2TQjZ+Q660/5yyROwQUDCAwVlMhLqDnfL5IZbNjak2YqPC6WKCdy8zaBYx
+xkJKftawZYqZ3noc4WqX30AKik/QYt04LpHWKw21EjqH1Bgq8T8yZlt264Ta21WR
+3pdqs59w+9g3BLnxUj599zmd4IUHuVWbuf5zhvn0jxi/FHFXBpDIqrsk3+LDiSKH
+citnP+Lq2ovFnR6IymwLMSLs6S4GNzM8b/G8tRdsoPOTVoFhYytEmyEtp84AAANw
+gzYvnYM2L50AAAAHc3NoLWRzcwAAAQEAk8EQralcwG+ePvHTuR5yKSOnP8cXpiWk
+1kiJJJKk+nD0Fp2DGE/KvqM8Xrppd/alke/xIyRBss7gjORN4LZhAdnybapg7azU
+vt1tvFgXLM/3ERa9DsaIRzmk2Opzh510yMBFDHJV2l9xanNXqzdtMBkrr6lokLqE
+i6OeiidOfw5oTaxXX9o50ttJN0nyqgn4/cE1IfVxQUWx92j/9Ymg140qIIP/OjfW
+4MI/l5A3JQckjTiqeMR+sqLgLF82onDLcINiR8CyW6VTDq7V3XCaFKavNJk7zI9W
+rCVobsP7Ajs2reRvdshN3oyTzDSu2cZqX/fCg0A4uTbXTiQqqjKtIwAAABUAuvme
+WsOGfke9R79z7CvMs3ec3ZUAAAEACueD6PoHskUd7sqXCMOmeXvXZG7I2H3kWBN2
+x8AFUfdv3rV/LiiCVMSGgZs0mFqsaprWpsTXvlSFOTJRNxTgPl3KhvugC0u0ruhG
+9xlTRtlu2GOjqPLxGyH1O8mNxVs+a+vc48NtMDB/9fdTVcQ0AVosAy76Ri6WyUBT
+S+fMdiRQIgYJW+iX7Jt0glyaYPMLDoBq6b/P7Af3DvboFfvynb9wyxwyfB//qnT2
+s8BWrUssJtfgxMrmePGUCRrXH68e6kPGhG4kBmQlw5SK6wWNv9zld5+++H77LHPh
+jffGKXFYNyzmVuqzKRmgh5y5mfZLqWkhYxDLMiKvdex9P/h6UwAAAQA5zCQR0IBW
+sav7BuYqTNArfHFJOE7MOIo1Bcrs1nmmni7N6vDeW3WR/wR2aEEti/cxKFE11r4c
+nNV97Y1bwy/sOh63pJc2Olotk0I2fkOutP+cskTsEFAwgMFZTIS6g53y+SGWzY2p
+NmKjwuligncvM2gWMcZCSn7WsGWKmd56HOFql99ACopP0GLdOC6R1isNtRI6h9QY
+KvE/MmZbduuE2ttVkd6XarOfcPvYNwS58VI+ffc5neCFB7lVm7n+c4b59I8YvxRx
+VwaQyKq7JN/iw4kih3IrZz/i6tqLxZ0eiMpsCzEi7OkuBjczPG/xvLUXbKDzk1aB
+YWMrRJshLafOAAAAFAXeCB+0dGeTXFy7JeSg5y5mfxGeAAAAEGRzYS1rZXktMjAx
+NzA0MDgBAgMEBQYHCAkKCw==
+-----END OPENSSH PRIVATE KEY-----
