+++ test/com/trilead/ssh2/signature/dsa-testkey-unprotected.txt
@@ -1,0 +1,20 @@
+-----BEGIN DSA PRIVATE KEY-----
+MIIDPQIBAAKCAQEAk8EQralcwG+ePvHTuR5yKSOnP8cXpiWk1kiJJJKk+nD0Fp2D
+GE/KvqM8Xrppd/alke/xIyRBss7gjORN4LZhAdnybapg7azUvt1tvFgXLM/3ERa9
+DsaIRzmk2Opzh510yMBFDHJV2l9xanNXqzdtMBkrr6lokLqEi6OeiidOfw5oTaxX
+X9o50ttJN0nyqgn4/cE1IfVxQUWx92j/9Ymg140qIIP/OjfW4MI/l5A3JQckjTiq
+eMR+sqLgLF82onDLcINiR8CyW6VTDq7V3XCaFKavNJk7zI9WrCVobsP7Ajs2reRv
+dshN3oyTzDSu2cZqX/fCg0A4uTbXTiQqqjKtIwIVALr5nlrDhn5HvUe/c+wrzLN3
+nN2VAoIBAArng+j6B7JFHe7KlwjDpnl712RuyNh95FgTdsfABVH3b961fy4oglTE
+hoGbNJharGqa1qbE175UhTkyUTcU4D5dyob7oAtLtK7oRvcZU0bZbthjo6jy8Rsh
+9TvJjcVbPmvr3OPDbTAwf/X3U1XENAFaLAMu+kYulslAU0vnzHYkUCIGCVvol+yb
+dIJcmmDzCw6Aaum/z+wH9w726BX78p2/cMscMnwf/6p09rPAVq1LLCbX4MTK5njx
+lAka1x+vHupDxoRuJAZkJcOUiusFjb/c5Xefvvh++yxz4Y33xilxWDcs5lbqsykZ
+oIecuZn2S6lpIWMQyzIir3XsfT/4elMCggEAOcwkEdCAVrGr+wbmKkzQK3xxSThO
+zDiKNQXK7NZ5pp4uzerw3lt1kf8EdmhBLYv3MShRNda+HJzVfe2NW8Mv7Doet6SX
+NjpaLZNCNn5DrrT/nLJE7BBQMIDBWUyEuoOd8vkhls2NqTZio8LpYoJ3LzNoFjHG
+Qkp+1rBlipneehzhapffQAqKT9Bi3TgukdYrDbUSOofUGCrxPzJmW3brhNrbVZHe
+l2qzn3D72DcEufFSPn33OZ3ghQe5VZu5/nOG+fSPGL8UcVcGkMiquyTf4sOJIody
+K2c/4urai8WdHojKbAsxIuzpLgY3Mzxv8by1F2yg85NWgWFjK0SbIS2nzgIUBd4I
+H7R0Z5NcXLsl5KDnLmZ/EZ4=
+-----END DSA PRIVATE KEY-----
