+++ test/com/trilead/ssh2/signature/ecdsa-sha2-nistp256-testkey-unprotected-newformat.txt
@@ -1,0 +1,9 @@
+-----BEGIN OPENSSH PRIVATE KEY-----
+b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAAAaAAAABNl
+Y2RzYS1zaGEyLW5pc3RwMjU2AAAACG5pc3RwMjU2AAAAQQQQX3TPOKh4a/N4cJFB
+TUkUyS7yOgddIpdgsij/xyLNuXAZOIGsXlexQyTd3Krq1dbmj57XUBLZsnUPKkrC
+KOF+AAAAsJCs7bWQrO21AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAy
+NTYAAABBBBBfdM84qHhr83hwkUFNSRTJLvI6B10il2CyKP/HIs25cBk4gaxeV7FD
+JN3cqurV1uaPntdQEtmydQ8qSsIo4X4AAAAgbIPKXLQtxkzECfQw7syiMp4r4uST
+J8a69bijYEOm/SUAAAASZWNkc2Eta2V5LTIwMTcwNDEwAQIDBAUG
+-----END OPENSSH PRIVATE KEY-----
