+++ test/com/trilead/ssh2/signature/ecdsa-sha2-nistp256-testkey-unprotected.txt
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIGyDyly0LcZMxAn0MO7MojKeK+LkkyfGuvW4o2BDpv0loAoGCCqGSM49
+AwEHoUQDQgAEEF90zzioeGvzeHCRQU1JFMku8joHXSKXYLIo/8cizblwGTiBrF5X
+sUMk3dyq6tXW5o+e11AS2bJ1DypKwijhfg==
+-----END EC PRIVATE KEY-----
