+++ test/com/trilead/ssh2/signature/ecdsa-sha2-nistp384-testkey-unprotected-newformat.txt
@@ -1,0 +1,11 @@
+-----BEGIN OPENSSH PRIVATE KEY-----
+b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAAAiAAAABNl
+Y2RzYS1zaGEyLW5pc3RwMzg0AAAACG5pc3RwMzg0AAAAYQSHxVgeI3jxXqmbFJi+
+a/iHRtpWAvbFi5iTBAaWBuPQnX3l6LWVFXzD9NOnp4zOwkWhQKOWlzK2uxE8xoil
+MRHeCsDe6CnvapTNbLAS+QSmXbqnplq56pdycO8V7HIlZwkAAADgGvTURhr01EYA
+AAATZWNkc2Etc2hhMi1uaXN0cDM4NAAAAAhuaXN0cDM4NAAAAGEEh8VYHiN48V6p
+mxSYvmv4h0baVgL2xYuYkwQGlgbj0J195ei1lRV8w/TTp6eMzsJFoUCjlpcytrsR
+PMaIpTER3grA3ugp72qUzWywEvkEpl26p6ZaueqXcnDvFexyJWcJAAAAMEcPAUAa
+E4jEtr+7GgAwYQCuHfspoBS1cfTsuqjmDYiVcr2obG1NYKHRbLWjj5CqAgAAABJl
+Y2RzYS1rZXktMjAxNzA0MTABAgMEBQY=
+-----END OPENSSH PRIVATE KEY-----
