+++ test/com/trilead/ssh2/signature/ecdsa-sha2-nistp384-testkey-unprotected.txt
@@ -1,0 +1,6 @@
+-----BEGIN EC PRIVATE KEY-----
+MIGkAgEBBDBHDwFAGhOIxLa/uxoAMGEArh37KaAUtXH07Lqo5g2IlXK9qGxtTWCh
+0Wy1o4+QqgKgBwYFK4EEACKhZANiAASHxVgeI3jxXqmbFJi+a/iHRtpWAvbFi5iT
+BAaWBuPQnX3l6LWVFXzD9NOnp4zOwkWhQKOWlzK2uxE8xoilMRHeCsDe6CnvapTN
+bLAS+QSmXbqnplq56pdycO8V7HIlZwk=
+-----END EC PRIVATE KEY-----
