+++ test/com/trilead/ssh2/signature/rsa-testkey-unprotected-newformat.txt
@@ -1,0 +1,30 @@
+-----BEGIN OPENSSH PRIVATE KEY-----
+b3BlbnNzaC1rZXktdjEAAAAABG5vbmUAAAAEbm9uZQAAAAAAAAABAAABFQAAAAdz
+c2gtcnNhAAAAASUAAAEBAJCdQJNgQPpBmlratR9x70/h63ELD7itF1mjM60jyo+b
+RCCzvoi/fDA+L7RwBq9Clzje+lkWpP/A6Duw76ksWuulxrtvFf/TcvRaWndWZojS
+EqRG2O51yM+GMfxxjsnpkvli1/MewVri2gGDcLrd+ws3mH1wryn4eYrLW/93ZG/+
+GckQegLoWlwYajNCgA1BqjHBlUbjMz2tI2DuOf+9Y/VB8QJ/CB9y/nyjHAzGFOoP
+y1c9Vnk0qqREghAUyAuJt6PKOMt2J3SUrnaMldFNuTAygdS4Ttt/EQTlP8xvursF
+GOoR1olEmOIHCIMYDJwHrM+/zmhWhkS45WmhV9XZh1MAAAPQmOYRGZjmERkAAAAH
+c3NoLXJzYQAAAQEAkJ1Ak2BA+kGaWtq1H3HvT+HrcQsPuK0XWaMzrSPKj5tEILO+
+iL98MD4vtHAGr0KXON76WRak/8DoO7DvqSxa66XGu28V/9Ny9Fpad1ZmiNISpEbY
+7nXIz4Yx/HGOyemS+WLX8x7BWuLaAYNwut37CzeYfXCvKfh5istb/3dkb/4ZyRB6
+AuhaXBhqM0KADUGqMcGVRuMzPa0jYO45/71j9UHxAn8IH3L+fKMcDMYU6g/LVz1W
+eTSqpESCEBTIC4m3o8o4y3YndJSudoyV0U25MDKB1LhO238RBOU/zG+6uwUY6hHW
+iUSY4gcIgxgMnAesz7/OaFaGRLjlaaFX1dmHUwAAAAElAAABABOK3zaCoP85TDXK
+hy3DSdpczMoWQGUQeMbsinFX3RpTQI7MLoEuohRbbjr6YcSoIkX0nl8Q5d1fQfo6
+gUBf8JxigqO1EM+LRuLAGhAgbriLF0ahykm848kES/GuexRd1ZdSi+KVdBMzaZGA
+dwR+3LxTnvwx0ntLF1fx96zmnt6zbN8YepMpUj3TyI5u7AZyxENR4rI/xHFve4Yy
+Ijp38U1L4a3vjd1RFkdAceoC46i38vyoe/PkyFmIYDkehosUJp6PKv8yHNI6HZ2c
+M0d9pokOSNzke5g524YeeecRJggDFGtkIi/RCiH00D3m6DH3AzymSpFoivSArHLP
+wPzl5p0AAACBAKLLjhPtAe6RH6n7ZGdLHKvi9gpD7Z9ZuNce69dSiJfqrOmqGtW5
+2RR5ouPFjMOrZlf81jHNi5Z+cpQX/1n0ym94Jb9q1mqR2ksVURnopcxASVlES4Z0
+KHGhEm574640iut/HZfDCjjkzko92+FGL5f8ADZ3Hx0JfDqP4b4yq2AxAAAAgQD5
+JBMhcmipIcU6YHeAjyG59EPUgQQdOGin8C9Q3VWlZ2NmRAtn/yVB4rylz8ONYfMO
+RVWF2eUGTVl94ylXuWVSKDDvninn9hB1bdXd4uzax3NFnJX2XWaf3FJmaUwnLowZ
+yC0yzt1kc/EX/2wGd/Hpan6cdFiNUEf6Ynw7CG7xSwAAAIEAlJh7ZxxN6j2aY7TJ
+YBs1ELG5Zmugzr3YT8LlhXG8+B6tBLdTuYdcF73WXnlTjHRlHV3V1Pjl2cBgPB/W
+WYAo4UypCwAMnWrWKq2ZPHbyQT9Vh28hLErSskgXubU+RurImzpmDSVcT9k1hR//
+T+WLeD4MPVmR9K/r7ui9VX34hRkAAAAQcnNhLWtleS0yMDE3MDQwOQECAwQFBgcI
+CQoLDA==
+-----END OPENSSH PRIVATE KEY-----
