+++ test/com/trilead/ssh2/signature/rsa-testkey-unprotected.txt
@@ -1,0 +1,27 @@
+-----BEGIN RSA PRIVATE KEY-----
+MIIEoQIBAAKCAQEAkJ1Ak2BA+kGaWtq1H3HvT+HrcQsPuK0XWaMzrSPKj5tEILO+
+iL98MD4vtHAGr0KXON76WRak/8DoO7DvqSxa66XGu28V/9Ny9Fpad1ZmiNISpEbY
+7nXIz4Yx/HGOyemS+WLX8x7BWuLaAYNwut37CzeYfXCvKfh5istb/3dkb/4ZyRB6
+AuhaXBhqM0KADUGqMcGVRuMzPa0jYO45/71j9UHxAn8IH3L+fKMcDMYU6g/LVz1W
+eTSqpESCEBTIC4m3o8o4y3YndJSudoyV0U25MDKB1LhO238RBOU/zG+6uwUY6hHW
+iUSY4gcIgxgMnAesz7/OaFaGRLjlaaFX1dmHUwIBJQKCAQATit82gqD/OUw1yoct
+w0naXMzKFkBlEHjG7IpxV90aU0COzC6BLqIUW246+mHEqCJF9J5fEOXdX0H6OoFA
+X/CcYoKjtRDPi0biwBoQIG64ixdGocpJvOPJBEvxrnsUXdWXUovilXQTM2mRgHcE
+fty8U578MdJ7SxdX8fes5p7es2zfGHqTKVI908iObuwGcsRDUeKyP8Rxb3uGMiI6
+d/FNS+Gt743dURZHQHHqAuOot/L8qHvz5MhZiGA5HoaLFCaejyr/MhzSOh2dnDNH
+faaJDkjc5HuYOduGHnnnESYIAxRrZCIv0Qoh9NA95ugx9wM8pkqRaIr0gKxyz8D8
+5eadAoGBAPkkEyFyaKkhxTpgd4CPIbn0Q9SBBB04aKfwL1DdVaVnY2ZEC2f/JUHi
+vKXPw41h8w5FVYXZ5QZNWX3jKVe5ZVIoMO+eKef2EHVt1d3i7NrHc0WclfZdZp/c
+UmZpTCcujBnILTLO3WRz8Rf/bAZ38elqfpx0WI1QR/pifDsIbvFLAoGBAJSYe2cc
+Teo9mmO0yWAbNRCxuWZroM692E/C5YVxvPgerQS3U7mHXBe91l55U4x0ZR1d1dT4
+5dnAYDwf1lmAKOFMqQsADJ1q1iqtmTx28kE/VYdvISxK0rJIF7m1PkbqyJs6Zg0l
+XE/ZNYUf/0/li3g+DD1ZkfSv6+7ovVV9+IUZAoGAUM1t/QKCzxHfGdoY62yw/goI
+KT6ZjO+zPWKnj9kU27LUISrTRFK/+bFR8JZo75Vjlez5JH4FFsz6bgR8HHN6zomT
+VKH/wNou2fob2UKtrr06JGoizHFRt05f6dz2GoujD0fXTr+hv7bmaKXd2JWakORE
+vR7QnIi9ZfZteu36d8UCgYA8PdEpzTRYCyLqJrKA6G91eHuYOXiLKl6c4E82GVpy
+bU0Pv/GCkNJO0GupwnvRKESrCl2bh3/G+vaU6k/6xUEBZESH60NiaZUYOIo7G3b3
+22fc9bOAqLZIRr2Co2xGQ4HCY8iB3ra4llP3swaj2ZloGSeOfkIXFuMbqnZFQOg1
+9QKBgQCiy44T7QHukR+p+2RnSxyr4vYKQ+2fWbjXHuvXUoiX6qzpqhrVudkUeaLj
+xYzDq2ZX/NYxzYuWfnKUF/9Z9MpveCW/atZqkdpLFVEZ6KXMQElZREuGdChxoRJu
+e+OuNIrrfx2Xwwo45M5KPdvhRi+X/AA2dx8dCXw6j+G+MqtgMQ==
+-----END RSA PRIVATE KEY-----
